@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 
 import App from './App.vue';
 import './registerServiceWorker';
@@ -7,9 +7,4 @@ import router from './router';
 // Main app stylesheet.
 import './scss/main.scss';
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  render: (h) => h(App),
-}).$mount('#app');
+createApp(App).use(router).mount('#app');
